gcloud/functions/design-tokens-api/requirements.txt
functions-framework>=3.5.0
google-cloud-storage>=2.14.0
